+++ src/react/src/component/index.ts
@@ -2,3 +2,4 @@ export * from './NotificationButton';
 export * from './MokPopup';
 export * from './Mok';
 export * from './MokAnalytics';
+export * from './CheckListPopup';